requirements.txt
Flask-SQLAlchemy == 2.4.3
psycopg2-binary ==  2.8.5
click == 7.1.2
six == 1.10.0
